requirements.txt
openai>=2.3.0
python-docx>=1.2.0
pyyaml>=6.0.3
tqdm>=4.67.1
lxml>=5.0.0
rich>=13.0.0
hypothesis>=6.0.0
aiohttp>=3.9.0
